Delete Course 'Exceeds' Criteria

+++ routes.js
@@ -179,14 +179,17 @@ router.put('/courses/:id', authenticateUser, asyncHandler (async(req,res,next)=>
 router.delete('/courses/:id', authenticateUser, asyncHandler (async(req,res)=>{ 
   const course = await Course.findByPk(req.params.id);
   const owner = req.currentUser;
-  console.log(owner);
-  if(course){
+  const ownerId = owner.id;
+  const courseOwner = course.userId;
+  if(course && ownerId === courseOwner){
     await course.destroy({
       where: {
         id: req.params.id
       }
     });
     res.status(204).end();
+  } else if (course && ownerId !== courseOwner){
+    res.status(404).json({message: "User must be Course Owner in order to Delete Course"})
   } else {
     res.status(404).json({message: "Course Not Found."})
   }
